Openemr fix telehealth signup locale (#6413)
* Add save and locale telehealth check instructions

Added a check for the locale settings which should be setup to avoid
clinic problems.

Added instructions that the user needs to log out and log back in after
saving the telehealth settings.

* Fix psr problems.

* Fix telehealth subscription signup.

Include instructions and the subscription id for the user to use when
they signup for the credentials.

Add a copy button and switch the alert of the subscription id to
actually show up in a success message info box.  Before the change the
user had limited understanding on how to proceed to the next steps in
the configuration system.

* Improve signup verbiage, add subscription id

To make the signup process easier for users, we break the signup into
three steps with instructions.  We also add saving the paypal
subscription id so we can check if the paypal process has already
completed and been signed up.

We show progress to the user as they signup.

* Fix if conditions, paypal key

Fixed the paypal key coming back from Sherwin to be the right key value.
Fixed the conditions on the subscription id.

* Include the free verbiage in the description.

* Updated telehealth plan for 14 day free trial.

* Defaulting values to avoid sql error.