Openemr fix 6574 6575 6576 fhir updates (#6577)
* Fixes #6575 refactor oauth2 client save

Made it possible to save an oauth2 client using the client repository.
This makes it so OpenEMR callers don't need to include the RestConfig
class or deal with the AuthorizationController to generate a smart app
client record if needed.  Module writers can register their own smart
app as needed using this functionality.

* Fixes #6576 FHIR Capability statement createUpdate

made it so the create update flag is set to false so API consumers know
they can't provider their own logical ids when creating a resource.

Fixes #6576

* Add systems for task/questionnaire

IN preparation for future Task and Questionnaire endpoints we want to
add the system constants here.

* Add mapped service helper methods.

Add helper methods for grabbing a token using a subset of codes as well
as grabbing a service based on a specific code.

* Fix url bug in utils service to prevent slashes

* Add helper method to Token search for isUuid

* Fixes #6574 QuestionnaireResponse search

The questionnaire response search would fail when searching on things
such as id due to duplicate column names.  In order to make this service
work in a FHIR context we need to fix up the search method and handle
the uuid translations.

Fixes #6574