fix: Update patient_tracker.php (#6595)
* Update patient_tracker.php

fixes php8 error not showing up in the logs - related to the implode function blocking FlowBoard from proceeding when a patient replies "CALL" indicating they need a callback regarding an appointment reminder they received.

It also removes the ability to add a note about the Callback request.  Instead it is replaced with an icon to directly open the SMS Bot so staff/front desk can directly reply to the patient via SMS, if SMS is allowed and they have a mobile/cell phone listed in demographics.  If not, just the CALL icon is displayed.

* styling

* add attr_js from review

---------

Co-authored-by: stephen waite <stephen.waite@cmsvt.com>