feat: Recent patient list (#6641)
Introduce a Recent Patient list which is updated when the demographics page loads and sets a PID. The number of patients on the list can be set in Admin > Config > Appearance and the columns to render can be customized with the Recent Patient Columns List. The default UI is a tab on the Patient Finder window.

Developers can programmatically get the list by using the new `getRecentPatientList()` function in PatientService and can update the list by calling `touchRecentPatientList()` and passing in an array of patient data.

Several Twig changes are incorporated as side effects. First, the introduction of a new Twig global variable called `session` which is a reference to `$_SESSION` Second, an update to `csrfToken` function which allows the template to pass a field name to as a parameter, the default remains `_token`. Lastly, a new function called `aclCore` was added which is a wrapper to `AclMain::aclCheckCore()` with the same parameter and return signature.