fix: Problem list UI improvements (#6661)
Allergies, Problem List, and Medications now have a max-height to allow for card-based scrolling to accommodate long lists. The one caveat is any allergy with a severity of severe, life threatening, or fatal will render as a badge on the top of the allergy card and will never be subject to scrolling. Non critical allergies, all problems, and all medications are displayed as a list-group-item.