feat: openemr 6752 6751 address book user rest api (#6767)
* Fixes #6752 address book config setting

Allows an address book config data type to be added for the global config.

Fixes #6752

* Fixes #6751 add user endpoint

Adds an endpoint for retrieving users from the system.  If there are
users we wish to retrieve that are not practitioners (such as external
entities that are in the address book), we can't use the practitioner
api.  Instead we expose the users endpoint.

Fixes #6751

* Left popup extension out in addrbook

* Fix use clauses, attribute escaping.

* Switch to relative path on api, pathing fix.

After conversations with admins we decided to go with a relative api url
for the internal api usage instead of the oauth2 full path.

Also fixed a path issue.