Fixes #7319 ccda extra drug_unit list options (#7321)
* Fixes #7319 ccda extra drug_unit list options

This fixes the error in the ccda that was creating tons of entries in
the drug_entry list options list.  The CCDA was not checking against
empty values and would continue to create new list_options for each
immunization / prescription drug_unit when the unit was an empty string

I also added some debugging instruments on the prescription and
immunization process which makes it easy to compare for a ccda what is
being parsed and what is being stored in the database.

Added a debug flag to all of the ccda commands so that these options are
available for people needing to do any debugging of the process.

* Exclude 0 values from amount administered checks.

We decided that we wanted to exclude recording 0 values so updated the
conditions on the immunization and prescription to account for that.