Fixes #7323 Billing Code Widget,deprecate isOption (#7324)
Fixes #7323 
This is a billing code selector widget that was used in options.inc.php.
I extracted it into its own class file so I can use it in a module as
well as use it in options.inc.php.

I also moved the options.inc.php isOption command to the LayoutUtils so
it can also be used externally.  I marked the isOption function as
deprecated.