Fixes #7503 user admin create empty google sign in (#7504)
* Fixes #7503 user admin create empty google sign in

Fixes #7503 by allowing the multiple users to have an empty google sign
in email address. If the address is empty it sets the value to null
which doesn't trigger the unique constraint.

* Fixes unit tests.

Was missing the null value on the insert, thought the custom escape
would handle it... but no.  Tried to refactor to be bound parameters but
found out the insert happens inside AuthUtils... instead of breaking the
heavily vetted password creation pieces by changing up the function
logic, backed out all the code and made the simple fix... sad.