Feat openemr fix 7480 7494 email prescription (#7495)
* Prep template to be moved to twig

* #7480 #7494 email prescriptions

This supports sending a prescription as a
pdf to an email address, sending the prescription inline as a formatted
html email, and sending the prescription in the native mail client using
the mailto: protocol.  Mailto: doesn't support email attachments so the
pdf option currently doesn't work with this.

When the Default Mail Client is selected (#7494) then the mailto link is
generated and if the browser or operating system has a mailto: protocol
handler setup it will launch the registered handler with the subject and
email body populated.

* Initial refactor to make edit page extensible

* Fixes #7496 move erx general edit to twig

Made it so the edit prescription dialog is in twig instead of in the old
html format.  This fixes #7496.  I've done some testing but probably
need to do more testing to verify we've handled all the use cases.

* Form save button position, bug fixes

Fixed a bug in the eye base where the window closing wasn't working
properly.

Moved the prescription action buttons out of the demographics.php file
and put them directly into the prescription dialog forms.

This helps fix some UX buttons flow issues as well as allows us to move
the buttons around as needed in the form.

Added a feature to have the form buttons on top of the form or on the
bottom.  Tried to have an option to show both top and bottom but that
fails in demographics.  I put the option in the demographics as well as
in the prescriptions form.  Hopefully other forms will follow suite and
allow the customization.

* Fix escaping issue.

* Not sure how this got left out but it did.

* Fixes #7525, #7524 prescription faxing

Made it so the prescription pdf faxing button will ask you to continue
printing if the signature isn't setup properly (at least to let users
know about it). Fixes #7525.

Fixed some php8 typecast errors preventing the pdf signature from
printing in the print to fax button for prescriptions. #7525.

* Switch to promise for erx dialog

* Handling variable escaping.

* Mark deprecated function