Fixes #7646 portal update page title tag (#7647)
* Fixes #7646 portal update page title tag

Made it so the patient portal will update the page title tag whenever
the single page app navigates to a new 'page'. I do this whenever the
page is persisted.

If the user comes back to the dashboard page it reverts to the default
page title set in the home.php file.

Fixes #7646

* Forgot about page refresh title

Had to update the page so we set the page title if the user refreshes
the screen.