Bug openemr #7276 snomed lookup fixes (#8033)
* Improved code lookup for snomeds and large tables

Changed up the search mechanism for code types with external tables that
define a fulltext search on the search term.  Use the database search
features for improved lookup performance. Made this generic so it uses
all the same code lookup functionalities instead of specific to snomed.

Also made it so tables can define skipping the table count properties in
the find code files (find_code_dynamic and select_codes).

Made it so that ICD10 and other tables show a list of
codes when the finder first pops up, whereas snomed when it first shows
will show an empty table with a message that you need to put in a search term

I brought back the auto-complete typing though I bumped it up to 350ms
delay instead of 200ms delay.  This fixes the enter key closing the
dialog bug as well as makes the finder more usable.  With the improved
search performance the autocomplete now can function well.  Will have to
test this when there's a lot of concurrency.

Fixes #7276

* Fix PSR styles

* Remove code_set_search_sct and unused comment

Not sure how my merge tool still included the code_set_search_sct method
even though I explicitly told it to not include it on the cherry-pick.
Either way, out it goes.

Also removed minor unused comment line.

* Fix table indexes for snomed ct codes

This speeds up the snomed lookup dramatically.

* Remove double spaces in files.