feat: Update eye_mag_functions.php to refine Imp/Plan Builder search (#8093)
* Update eye_mag_functions.php

This PR refine the Imp/Plan Builder to search for whole words/phrases in a clinical field.  For example, prior code allowed positive hits for "hole" when the clinical field included "whole" and "wholey" etc.  This Limits it to "hole" by addition a regex to add \b$term\b to the logic.  It also limits searches to fields with activity=1.

* styling

---------

Co-authored-by: stephen waite <stephen.waite@cmsvt.com>