Openemr feat 8162 8163 clinical notes forms (#8246)
* WIP clinical notes form

Don't want to lose track of this. Implemented twigify version of
clinical notes.  Moved js logic to clinical-notes.js file.

Also made it so default type/category are selected in notes.
Option to hide type/category if list options are made inactive.

Fixes #8162
Fixes #8163

* Reverting items for debug, formatting

Formatting was fixed in a different PR.  Reverting config change.

Also removing debugger statement and console.log from formatting.js

* Clinical notes dropdown defaults, codetext display

This fixes issues when cloning / copying the row and default category
not being set.

Fixed up issues in the twig template dealing with the default value.

Also changed up the report display so it always shows the code column
which helps with the formatting display a bit.  Shows unspecified if
there is no value populated.

Fixes #8163 Set the default value for type and category
Fixes #8245 Made the code text display

* Fix initial default value, js bug when lists empty

Had a bug where if there are no active types or categories the data does
not save.

Also fixed a problem where if there are no default list items it selects
the first item in the list.  Users expected behavior is that the
dropdowns would be empty.

* Remove leftover comment