Automate Inferno Testing (#8409)
* ci: automatic inferno testing

* update onc-certification-g10-test-kit

* update onc

* tweak inferno compose

* fix(openemr.sh): correct working directory

* DB credential fix, unit tests against inferno

Fixed the db credentials for OpenEMR to connect to mysql using the
environment variables.  This makes the openemr-cmd snapshot installation
work as long as the auto_configure.php file is there.  Still need to
work on getting the auto_configure.php to be copied into the docker
container.

Fixed up the submodule to enable CORS on the nginx installation.

Wrote out a preliminary api test for the single patient api.  Found out
the us_core_311 suite is very different in its terminology support
compared to the g10_certification.  So tests passing in
g10_certification are failing in the us_core_311 which is very
frustrating.

I can get individual resource results on the us_core_311 suite whereas
on the g10_certification the most granular I can get to right now is the
entire single patient test suite (all 300+ tests).  I need to either
figure out a way to handle the reporting of the entire test suite in a
way that is easily visible from the phpunit test runner or figure out a
way to make the resource tests executable as individual groups.

I also discovered a bug in the redis docker installation where it fails
to change the owner to be redis:redis on the /data folder which causes
inferno test session instances to fail to create.  It looks like the
.keep file in the folder directory causes redis to skip the permission
change.  To fix it manually you have to run the following command which
I found on the redis docker instructions page.

docker run --rm -v /your/host/path:/data redis chown -R redis:redis /data

* Inferno test suite ci engine

Got the system to build and run the test on the first invocation of
run.sh.

Tearing down the dockers and re-running run.sh fails on the redis server
with the permission changes.  Redis leaves the data directories as a
999:999 uid:guid which the current user doesn't have permissions to copy
the files over and the script fails. So far don't have a solution that
fixes it outside of just removing the
onc-certification-g10-test-kit/data/redis folder after tearing down the
dockers.

It also looks like the test suites fail at times due to db contention
when the test is running and when I query for the test results in the
phpunit test suite.  I need to figure out if there is a way to run it
w/o contention.

Next steps are to figure out a ci workfile and then to also make sure
the coverage engine is turned on.

* Inferno workfile that I'm testing.

Going to commit and try a clean install to see if I can reproduce the
docker build failing.

* Fix openemr-cmd implementation

Got the openemr-cmd to work properly by downloading it as part of the
CI.

Renamed tests to indicate the 311 version so we can add in the 700
version as needed.

* Passing 3.1.1 G10 certification single api test

Got the TLS finally disabled and working

* Initial coverage support.

Added coverage support to the dispatch.php file that gets called on
every inferno request.  It only runs if the environment variable is set
so we can collect coverage information.

The coverage process is extremely slow taking 30 seconds per http
request from inferno which means it will take around ~2 hours for the
coverage report to run for inferno.  Supposedly the pcov extension will
take a lot less time so going to see if I can experiment with that and
see how that works.

* Fixes to help coverage complete

* Style fixes, fix refresh token auth

* More style checks, fix refresh token type

Client api registration is symmetric key not assymetric for using
refresh token.

* Removing coverage

Coverage is taking 2 hours to execute.  Disabling for now.

* Implemented bulk fhir api test

Abstracted the credentials assert generator into its own class so it can
be used for testing.
Implemented a test for bulk patient export via fhir.

Pulled out the single patient and bulk patient pieces into its own trait
so we can reuse it.

* Fix auto_configure.php by curling it to the root

This fixes the ci dockers not having the auto_configure setup.

* PSR style fixes

* Change to string on YAML file

It shouldn't matter whether its quoted or not, but going to exactly
follow the github action checkout guide and see if that impacts things.
Otherwise we have to look at using a Personal Access Token (PAT) or
something else to get the private modules to checkout.

* Use github personal access token for repo access

Going to attempt to use a PAT as a secret environment variable and see
if this can resolve our access permissions.

* Removes the upload of test results

There are  no result files to upload outside of whats in the phpunit
test.  So removing that from the workflow file.

* Use ubuntu-latest for deploy image

---------

Co-authored-by: Stephen Nielson <snielson@discoverandchange.com>