Openemr fix 8846 questionnaire us core (#8877)
* implement #8846 questionnaire endpoints

Implements the basic questionnaire endpoints.  Brought in the
implementation from my Discover and Change assessment module and had to
modify a number of things to make it compatible in core.

Brought in some of the HL7 questionnaires that we can use for testing.
Wrote out test cases so we can verify the services and rest endpoints
are all working properly.  Test cases can definitely be added as there
are a lot of branch conditions in the QuestionnaireService and
QuestionnaireResponseService classes but tried to go for a basic smoke
screen test for list and get operations.

Next step is to make sure everything is compliant with US Core 8.0
profile.

* FHIR _questionnaire implementation and tests

Implemented the extended _questionnaire attribute via the new class
OpenEMRFhirQuestionnaireResponse.

Also implemented an integration test for verifying everything conforms
to Us Core.

* Fix phpstan issues and psr styling

* Fix service class constructors.

* Fix code review issues, rector problems

* Great... missed one.