Openemr fix #8801 observation form us core (#8889)
* Refactor observation files into MVC architecture

Implemented a Controller for the new,report,save views of the
observation form.  Implemented a service class to put observation
retrieval and saving into.

Wrote unit tests to ensure the basic functionality works before we add
additional data elements to the observations.

* Fixing code review issues, phpstan issues

Added AI generation comments, fixed phpstan complaints.

* Remove unused tests, refactor queries

Refactored the queries to use QueryUtils and changed tests that had
little utility in order to mark them incomplete for the future.

* WIP updated observation form gui for #8801

Updated the observation form GUI to have a list view of the current
observations and an edit view of the observation with the ability to
have sub-observations and stubbed out GUI for adding in questionnaire
responses.

Added ability for code description to be editable in case of adding your
own codes.

Added all of the new observation types needed for USCDI simple
observation conformity.

Added observation status field to allow status to be set.  Defaults to
preliminary status.

Date defaults to current date when adding an observation.

Added ability to have sub-observations to handle cases where you have a
panel observation.

Added search field on observations.  It is still very rudimentary and I
may take it out as it just handles it client side which isn't helpful
for dealing with lots of observation data.

Added in some basic form validation.

Removed the legacy syndrmoic surveillance hard-coded LOINC code types
which had no explanation of what they did.

Refactored HTML code to use twig.

Removed old html files

* Update controller unit tests, minor todo

Add a minor todo on abstract observation service to move code into
ObservationService

* Observation saving, questionnaire linking

Got the observations saving and questionnaires linking properly.  Fixed
up navigation flow between new, edit, and list.  Fixed up the reports to
show more of the actual observation details.

Did prep work for sub-observations which is my next task list.

Commented out a number of search and filtering functionality that can be
added in later, but for now, just trying to focus on the core
requirements needed for US Core.

Fixed an issue with the fhir questionnaire endpoints making them not
work properly.

Implemented the questoinnaire search and linking for an observation in
order to handle the observation justification.

* Sub observation saving

Got the sub observation records saving in the service as well as in the
controller and twig.  Fixed some issues where sub-observations were not
displaying properly.

* Observation top level deletion

Got the top level deletion working for observation.  In reality we set
the activity flag to 0 on the observation instead of whole-scale
deleting them.

I now need to handle the sub-observation deleting which will be the
intersection between the new and old on the postData.

* Sub observation deletion

Got the sub observations deleting properly.

* Fix unit tests, phpstan, styles

* Fixed the semicolon issue

* Unit tests for CodeTypesService::collectCodeTypes

implemented unit tests before refactoring the class to deal with the
code review suggestions.

* Refactor collectCodeTypes method to use match expr

Implementing code review to refactor collectCodeTypes method in
CodeTypesService to use match instead of if/else clauses.

* Fix phpstan issue

* Fix test cases on fresh database

Integration test relied on existing db data.  Had to fix it so it works
off a fresh database

* Fix styles and rector issues.

Fixed the styles and the rector issues.

* Fix services unit test

Had pre-existing data for the form_id

* global state sucks... fix the unit test

Super frustrating that we are relying on global state here but until we
can get away with it with a Service container or some other form of DI
we get to live with it.

* Fix more rector issues

* Improve collectCodeTypes mapping function

* Fix db issues, escaping, code type exception

Fixed up db issues where we were missing indexes.

Also fixed up escaping issues in the twig templates.

Finally fixed the code types service to throw an exception on the
collectCodeTypes method if the category is invalid.

* Fix tests