Fixes #8958 oauth2 session (#8961)
* Fixes #8958 oauth2 session

The OAuth2 Authorization Code grant flow was not working properly due to
changes with the session handling / management.  Needed to fix the issue
with the session being cleaned up on every request if it wasn't a local
api request (launched within the GUI application in OpenEMR).

Also discovered that the launch/patient scope was not working with the
patient selector due to the CSRF token not properly populating due to
the session changes so fixed that issue also.

Fixes #8958

* Fix session bug with unit tests

The unit tests were mocking the controller session but not the request
session which ended up throwing an exception.  This actually revealed a
bug in the exception handling code which is now fixed.