Feat openemr #9131 patient profile (#9154)
* WIP patient data us core elements stubbed out

* Implement US Core 8.0 Patient Profile #9131

This implements the US Core Patient Profile
Adds two new lists
 - administrative_sex which holds the list options for the new
   patient_data.sex_identified column
 - yes_no_unknown which holds the list options for the patient_data.

Renames on Demographics New/Edit page the following fields
- Sex -> Birth Sex
- Interpreter -> Interpreter Comments

Adds the following Demographics New/Edit page fields
- Sex (mapped to administrative_sex list_id)
- Interpreter (mapped to yes_no_unknown list_id)

Adds in the deceasedDate logic.

Fixes #9131

* US Core Profile Versioning for breaking changes

Patient US Core 8.0.0 profile has breaking changes on us-core-sex
extension between US Core 6.1.0/7.0.0 and US Core 8.0.0.  We need to be
able to differentiate between these options.

We need to set the highest supported profile version that is supported
when there are breaking changes.  Most of US Core has been forwards
compatible (they only add fields / properties) but Patient is the first
breaking change I've run across.  The ground work is laid to have
entirely different api routes based on the version (just set the highest
supported version per route definition file) but for now we set it as a
global flag of what the maximum supported version will be.

In this case, setting the supported version to 7.0.0 uses the FHIR Code
data type for Patient whereas choosing version 8.0.0 uses the FHIR
Coding datatype for us-core-sex extension.  This adjusts what values get
reported in the getProfileURIs methods based upon what version is
supported.

If we end up on some of the other resources having to deal with breaking
changes we'll roll out complete end points for compatibility support.

* Fix fatal error and extension not added

* Fix system on unknown interpreter

System should have been a URI not a code type

* Swap display / system parameter error

* Fix date deceased time stamp

* Fix deceased date single entry requirement

The deceased property is a 0..1 requirement so it is either a boolean
(which in this case is false) or it is the date of death for the
patient which we track.

* Fix styles

* Fix missing class definition

* US Core 8.0.0 Profile Tests, FHIR Write parsing

Implemented a bunch of the parsing needed for FHIR write to have
compatability with US Core 8.0.0.

Implemented a unit test file to make sure we are complying with US Core
7.0.0 and US Core 8.0.0 patient profiles.

* Left out Sex field in new db setup

* Fix rector issues, error w/ globals refactor

* Fix e2e tests with new sex_identified field

* Fixed location of where list_options are created