Feat openemr 9075 average blood pressure (#9104)
* WIP average blood pressure

Implement the foundational framework for average blood pressure by
allowing for calculated vitals records that are computed as a derivative
to primary source vital calculations.

Right now this calculates 3day mean, last 5 vital readings, and mean of
all vitals in an encounter.

Still need to handle the aggregation of bps and bpd which is a weakness
with this current approach.  I may need to break the tables up into a
calculation and then have sub columns that are part of that calculation.

* Observation blood pressure sub-components

Fix calculation events not firing

Fixed calculation sub-components for average blood pressure.

Added interpretation property for FhirObservationTrait
Fixed a number of phpstan and rector issues.
Fixed a number of phpstan and rector issues.
Fixed a number of phpstan and rector issues.
Fixed a number of phpstan and rector issues.

* Fixing missing table syntax

* Fix missing description on code

* Fix broken rebase with bracket

* Fix rector and styles

* Remove foreign key constraints

* Add drop statements, cleanup comments