Feat openemr #9373 bulk fhir (#9384)
* Fix bulk failing resources w/ date and session

A number of resources that rely on session values were failing as the
session was not populating in the Bulk Export and in the Provenance
resource.

Fixed queries that were breaking with the _since parameter that search
based on last updated.

* Make it easier to generate api test clients.

* Fix route requests

* Rename interface to prep for globals DI

* Fix patient profile validation, bulk export

Got the patient profile validation working properly with injection of
the globals in the controller and service classes.

Move IGlobalsAware to be a service interface instead of just a command
interface for OEGlobalsBag injection.

* MedicationRequest Fix quantity and reported[x]

* Fix proper access controls in FHIR doc api

* Fix MedicationAdherance

MedicationAdherence wasn't showing up for linked prescriptions so fixed
that issue.

* Rename class back to original

Not sure how but somehow the class got renamed, so fixed that.

* Fix _lastUpdated search for MedicationDispense

* Fix php pass by reference error

* Refactored encounter dropdown into class

Moved the encounter dropdown from the TwigExtension into its own class
so it can be used in both twig and in PHP.

* Add encounter to immunization list

* Fix broken prescriptions when updating

When updating a linked lists_medication record from an existing
prescription, the save was failing.

* Added Encounter,Location to FHIR Immunization

* Fixes #9382 RXCUI code for medication

Fixed a bug when inventory is used where the drug_code is not properly
populated.

Fixes #9382

* Support shorthand ndjson, fix invalid request err

Invalid request w/ outputFormat for BULK FHIR export was throwing a
fatal 500 operation outcome error.

Fixed it so its an bad request (400)  and now support the shorthand
ndjson format.

* Add support to specimen for bulk export support

* Fixes for Labs, Procedures, Specimen

Specimen.value[x] must be a numeric value and not a string.
Removed usedReference as it was breaking the specification.

* Fix job request delete should return 404

* Implement FHIR Location Service

* Fix vitals summary observation

* Fix rector issues

* Oops... putting this back

* Fix fhir diagnostic profile

We were not adding in the profiles to the diagnostic report resources
which failed bulk fhir profile validation.

Had old code in the clinical notes diagnostic profile that was adding in
sample records.

* Remove extraneous logging

There were extra logs of resource uuids and patient pids

* Fix FHIR pulse oximetry, and temperature

Resources were not conforming to profile

* Removed commented code from twig extension

* Fixes #9386 US Core Vitals Fixes

Fixes #9386 - issues with us core vitals.

Removes bp and pulse oximetry sub-components as observations and bundles
them into the correct component implementation.

* Fix observation laboratory for bulk export

Fixes #9387 adding in missing fields in fhir

* Fix rector issues

* Add missing specimen identifiers

Was missing fields needed for specimen

* Rector and psr-12 fixes

* Fix additional rector issue

* Fix RelatedPerson uuid not showing properly

* Fix profiles for DiagnosticReport

Capability statement was not showing the correct profiles

* Fix export CareTeam,ADI,Care Experience,Treatment

* Fix RelatedPerson use/system list items

* Fix system user generate access token

access token was being forced to the bulk export system user

* Fix Location ACL issues, add location search

Added location search and fixed up permissions missing from the session
by using explicit vars instead of global $_SESSION

* Fix PractitionerRole,Location session

* Fix oxygen saturation, provenance

Fixed provenance with Care Experience and Treatment Intervention
observations.

Fixed missing oxygen saturation

* Fix database encounter field for immunization

* Fixes #9409 handling practitioner for reported[x]

Fixes #9409 by have an option to say the record is not a primary record and
to choose a user from the address book. At some point when we disentangle
external organizations from the address book and put them in facility we can
handle external organizations in addition to practitioners. If a user is
chosen without an NPI it will show the Primary Business Entity as the
reported organization.

* Fix code search on observations

Fixed the code search with the way system codes work for observation
form, care experience, and treatment intervention

* Updated sql

* Improve CLI for SMART EHR-Launch

* Fixes #9368, #9423

Fixes #9368 patient context population for patient context restriction
Fixes #9423 fix token serialization / deserialization

The token was not deserializing in inferno due to the + character from
encryption so needed to base64 encode it.

Also left out the context population from the rest api refactor. Made it
so the patient context gets populated on the request

* Fixes #9428 us core capability statement

Add the us core server url for the capability statement
Fixes #9428

* Fixes #9247 practitioner address info

Adds in the practitioner address support for FHIR Practitioner and
PractitionerRole which gets these resources passing certification.

Fixes #9247

* Fixes #9424 questionnaire category dropdown

Added category dropdown for questionnaire category registration so we
can group the categories and report them in FHIR via their category
attribute.

* Fix rector and phpstan issues

* Fix phpbcf issues

* Fix broken unit tests

* Fixes #9429 add diagnostic performer

Added the ability to have practioners be non-OpenEMR users so that
diagnostic report references can reference them.  See the issue for more
details.

Fixes #9429

* Fix rector issues

* Fix Location telecom system to be url

Instead of using website as the system, we use url as the telecom
location element.

* Removed left in code.

Had an encrypted string that was from a specific OpenEMR instance left
in the code.

* Fix phpstan and rector

* Another test case fix

Had another case where inferno was passing but vanilla openemr was not