Bug openemr #9460 token introspection (#9463)
* Refactored introspection into sep controller

Fixed the expiration bug and refactored the introspection into a
separate controller for single responsibility principle and better
testability.

* FHIR Token Introspection Fixes #9460

Added issuer to access token entity in order to comply with SMART on
FHIR v2.2 specification.

Added iss,fhirUser, and fhir token context params to introspection for
access token.  (patient, smart_style_url,need_patient_banner, etc).

Changed up introspect response to comply with RFC 7662 by returning a
response with active set to false for invalid tokens.  In order to
comply with inferno I return a inactive response when the authentication
fails as well but I log the error so people can debug.  Had to update
all the unit tests to handle that.

Ended up changing the openid-configuration url to be .well-known in
order to comply with the spec (and also our documentation).

Fixes #9460

* Fix broken unit tests from refactor